% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_publishers.R
\name{cargar_publishers}
\alias{cargar_publishers}
\title{Request all available publishers from \url{https://datos.gob.es/}}
\usage{
cargar_publishers()
}
\value{
a \code{\link[tibble]{tibble}} with two columns: publisher_code and publishers
}
\description{
Request all available publishers from \url{https://datos.gob.es/}
}
\examples{

cargar_publishers()

}
