% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{data_list_correct}
\alias{data_list_correct}
\title{Check data_list is in correct formats}
\usage{
data_list_correct(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
When new checks come up, add them in
the same format: logical tests first
and then add them to the if statement
}
