% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_publishers.R
\name{datos_publisher}
\alias{datos_publisher}
\title{Request all available publishers from datos.gob.es}
\usage{
datos_publisher()
}
\value{
a \code{\link[tibble]{tibble}} with two columns: publisher_code and publishers
}
\description{
Request all available publishers from datos.gob.es
}
\examples{

datos_publisher()

}
