% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_funs.R
\name{explorar_keywords}
\alias{explorar_keywords}
\title{Explore datasets by keywords and publishers in \url{https://datos.gob.es/}}
\usage{
explorar_keywords(keyword, publisher)
}
\arguments{
\item{keyword}{A character string specifying a keyword to identify a data set. For example, 'vivienda'.}

\item{publisher}{A character string with the \strong{publisher code}. Should be only one publisher code.
See \code{\link{publishers_available}} for the permitted publisher codes.}
}
\value{
A \code{\link[tibble]{tibble}} containing the matched datasets.
}
\description{
Explore datasets by keywords and publishers in \url{https://datos.gob.es/}
}
\details{
\code{explorar_keywords} works only for searching for one keyword for a given publisher. For example,
'viviendas' for the Ayuntamiento of Barcelona. If there are no matches for a keyword-publisher combination,
\code{explorar_keywords} will raise an error stating that there are no matches.

\code{explorar_keywords} returns a data frame with the following columns:

\itemize{
\item description: a short description of each of the matched datasets in Spanish
 (Spanish is set as default if available, if not, then the first non-spanish language is chosen).
\item publisher: the entity that publishes the dataset. See \code{\link{cargar_publishers}} for all available publishers.
\item is_readable: whether that dataset is currently readable by \code{\link{cargar_datos}}.
See \code{\link{permitted formats}} for currently available formats
\item path_id: the end path that identifies that dataset in the \url{https://datos.gob.es/} API.
\item url: the complete url of the dataset in \url{https://datos.gob.es/}. Note that this URL is not the access URL to the dataset
but to the dataset's homepage in \url{https://datos.gob.es/}.
}

In most cases the user will need to narrow down their search because the result of \code{explorar_keywords}
will have too many datasets. Beware that for passing the result of this function to \code{\link{cargar_datos}} the final
data frame needs to be narrowed down to only one dataset (that is, 1 row) and the structure needs to be the same
as from the original output of \code{explorar_keywords} (same column names, in the same order). See examples below.
}
\examples{

\dontrun{

library(dplyr)

kw <- explorar_keywords("vivienda", "l01080193") # Ayuntamiento de Barcelona

kw

# Notice how we narrow down to only 1 dataset
dts <-
 kw \%>\%
 filter(grepl("Precios", description)) \%>\% # Narrow it down to only 1 dataset
 cargar_datos('ASCII')

# Notice that we had to specify the encoding because printing the dataset returns an error.
# If that happens to you, try figuring out the encoding with readr::guess_encoding(dts$data[[1]])
# and specify the most likely encoding in `cargar_datos`

dts$metadata

dts$data
}

}
\seealso{
\code{\link{cargar_datos}}
}
