% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_access_url}
\alias{extract_access_url}
\title{Extract access URL to the actual data from data_list}
\usage{
extract_access_url(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract access URL to the actual data from data_list
}
