% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_dataset_name}
\alias{extract_dataset_name}
\title{Extract the dataset names from data_list}
\usage{
extract_dataset_name(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
For example, elecciones2016.csv and elecciones2014.csv
}
