% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_description}
\alias{extract_description}
\title{Extract description from data_list}
\usage{
extract_description(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract description from data_list
}
