% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_endpath}
\alias{extract_endpath}
\title{Extract the end path of the dataset that directs to datos.gob.es from a data_list}
\usage{
extract_endpath(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract the end path of the dataset that directs to datos.gob.es from a data_list
}
