% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract keywords from data_list}
\usage{
extract_keywords(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract keywords from data_list
}
