% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_language}
\alias{extract_language}
\title{Extract access languages available from data_list}
\usage{
extract_language(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract access languages available from data_list
}
