% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract all metadata from a data_list}
\usage{
extract_metadata(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract all metadata from a data_list
}
