% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_modified_date}
\alias{extract_modified_date}
\title{Extract the date at which the data was modified from data_list}
\usage{
extract_modified_date(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
The date is currently exported as a string but
should be turned into a a date class
}
