% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_publisher}
\alias{extract_publisher}
\title{Extract the publisher of the dataset from data_list}
\usage{
extract_publisher(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract the publisher of the dataset from data_list
}
