% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_publisher_code}
\alias{extract_publisher_code}
\title{Extract the publisher code of the dataset from data_list}
\usage{
extract_publisher_code(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract the publisher code of the dataset from data_list
}
