% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_publisher_name}
\alias{extract_publisher_name}
\title{Extract the publisher name of the dataset from data_list}
\usage{
extract_publisher_name(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract the publisher name of the dataset from data_list
}
