% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_title}
\alias{extract_title}
\title{Extract title from data_list}
\usage{
extract_title(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract title from data_list
}
