% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_url}
\alias{extract_url}
\title{Extract URL from datos.gob.es from data_list}
\usage{
extract_url(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Extract URL from datos.gob.es from data_list
}
