% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components.R
\name{extract_url_format}
\alias{extract_url_format}
\title{Extract the format of the dataset from data_list}
\usage{
extract_url_format(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
For example, csv or xml
}
