% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_resp}
\alias{get_resp}
\title{Make GET requests with repeated trials}
\usage{
get_resp(url, attempts_left = 5, ...)
}
\arguments{
\item{url}{A url, preferably from the \code{path_*} functions}

\item{attempts_left}{Number of attempts of trying to request from the website}

\item{...}{Arguments passed to \code{\link[httr]{GET}}}
}
\description{
Make GET requests with repeated trials
}
