% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_resp_paginated}
\alias{get_resp_paginated}
\title{Make GET requests over several pages of an API}
\usage{
get_resp_paginated(ch_url, num_pages = 1, page = 0, ...)
}
\arguments{
\item{ch_url}{URL to request from, preferably from the \code{path_*} functions}

\item{num_pages}{Number of pages to request}

\item{page}{The page at which the request should being. This should rarely be used}

\item{...}{Arguments passed to \code{\link[httr]{GET}}}
}
\value{
the parsed JSON object as a list but inside the items
slots it contains all data lists obtained from the pages specified
in \code{num_pages}
}
\description{
Make GET requests over several pages of an API
}
