% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_publishers.R
\name{is_publisher_available}
\alias{is_publisher_available}
\title{Check if publisher is available in opendataes}
\usage{
is_publisher_available(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Check if publisher is available in opendataes
}
