% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_readable.R
\name{is_readable}
\alias{is_readable}
\title{Get prefered method to download data (if possible)}
\usage{
is_readable(data_list)
}
\arguments{
\item{data_list}{A data_list similar to resp$result$items[[1]] that contains information on a dataset}
}
\description{
Get prefered method to download data (if possible)
}
