% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_api.R
\name{make_url}
\alias{make_url}
\title{Build a custom url using the httr url class}
\usage{
make_url(path, param, ...)
}
\arguments{
\item{path}{the end path of the dataset of interest}

\item{param}{arguments for a query}

\item{...}{any other arguments to building the path correctly. See \code{modify_url}}
}
\description{
Build a custom url using the httr url class
}
