% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\name{openes_load_publishers}
\alias{openes_load_publishers}
\title{Request all available publishers from \url{https://datos.gob.es/}}
\usage{
openes_load_publishers()
}
\value{
a \code{\link[tibble]{tibble}} with two columns: publisher_code and publishers
}
\description{
Request all available publishers from \url{https://datos.gob.es/}
}
\examples{

\donttest{
openes_load_publishers()
}
}
