% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_begin_end_date}
\alias{path_begin_end_date}
\title{Build a url with a complete begin-end date/ prefix URL}
\usage{
path_begin_end_date(start_date, end_date, param = NULL, ...)
}
\arguments{
\item{start_date}{Start date in YYYYMMDD format}

\item{end_date}{End date in YYYYMMDD format}

\item{param}{Extra parameters to add to the url. For this function, this is
useless because there's not further paths to the distribution end point. Keeping
the argument for consistency}

\item{...}{Extra arguments passed to \code{\link[httr]{build_url}}}
}
\description{
Build a url with a complete begin-end date/ prefix URL
}
