% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_catalog}
\alias{path_catalog}
\title{Build a url with a complete catalog/ prefix URL}
\usage{
path_catalog(path, param = NULL, ...)
}
\arguments{
\item{path}{the end path of the dataset of interest}

\item{param}{arguments for a query}

\item{...}{any other arguments to building the path correctly. See \code{modify_url}}
}
\description{
Build a url with a complete catalog/ prefix URL
}
