% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_api.R
\name{path_dataset_id}
\alias{path_dataset_id}
\title{Build a url with an ID of a dataset}
\usage{
path_dataset_id(id, param = NULL, ...)
}
\arguments{
\item{id}{dataset id from datos.gob.es such as 'l01080193-numero-total-de-edificios-con-viviendas-segun-numero-de-plantas'}

\item{param}{Extra parameters to add to the url.}

\item{...}{Extra arguments passed to \code{\link[httr]{build_url}}}
}
\description{
Build a url with an ID of a dataset
}
