% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_datasets}
\alias{path_datasets}
\title{Build a url with a complete datasets/ prefix URL}
\usage{
path_datasets(param = NULL, ...)
}
\arguments{
\item{param}{Extra parameters to add to the url. For this function, this is
useless because there's not further paths to the dataset end point. Keeping
the argument for consistency}

\item{...}{Extra arguments passed to \code{\link[httr]{build_url}}}
}
\description{
Build a url with a complete datasets/ prefix URL
}
