% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_explore_keyword}
\alias{path_explore_keyword}
\title{Build a url to search for a given keyword in the datos.gob.es API}
\usage{
path_explore_keyword(keyword)
}
\arguments{
\item{keyword}{A string with the keyword to build the path with.}
}
\description{
Build a url to search for a given keyword in the datos.gob.es API
}
