% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_api.R
\name{path_publishers}
\alias{path_publishers}
\title{Build a url with a complete publishers/ prefix URL}
\usage{
path_publishers(param = NULL, ...)
}
\arguments{
\item{param}{Extra parameters to add to the url. For this function, this is
useless because there's not further paths to the publishers end point. Keeping
the argument for consistency}

\item{...}{Extra arguments passed to \code{\link[httr]{build_url}}}
}
\description{
Build a url with a complete publishers/ prefix URL
}
