% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{permitted_formats}
\alias{permitted_formats}
\title{Current readable formats from \url{https://datos.gob.es/}}
\usage{
permitted_formats
}
\description{
Current readable formats from \url{https://datos.gob.es/}
}
\examples{
permitted_formats
}
\keyword{datasets}
