% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\docType{data}
\name{publishers_available}
\alias{publishers_available}
\title{Available publishers that `opendataes` can read}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.}
\usage{
publishers_available
}
\value{
a \code{\link[tibble]{tibble}} with two columns: publishers and publisher_code
}
\description{
Available publishers that `opendataes` can read
}
\examples{

publishers_available
}
\seealso{
\code{\link{cargar_publishers}}
}
\keyword{datasets}
