% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\name{translate_publisher}
\alias{translate_publisher}
\title{Translate publisher code to publisher name}
\usage{
translate_publisher(code)
}
\arguments{
\item{code}{A publisher code}
}
\description{
Translate publisher code to publisher name
}
